# Conflicts:
#	backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/integrationtest/UserEndpointTest.java
#	backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/unittests/User/ApplicationUserRepositoryTest.java
#	backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/unittests/User/ApplicationUserServiceTest.java

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/UsersApi.java
@@ -150,7 +150,7 @@ public interface UsersApi {
         produces = { "application/json" }
     )
     default ResponseEntity<UserDto> usersIdGet(
-        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         getRequest().ifPresent(request -> {
             for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
